+++ src/main/java/com/dovendev/track/jpa/repositories/TrackRepository.java
@@ -4,5 +4,5 @@ import com.dovendev.track.jpa.entities.Track;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-public interface TrackRepository extends JpaRepository<Track, Integer> {
+public interface TrackRepository extends JpaRepository<Track, Long> {
 }
