+++ src/main/java/com/dovendev/track/jpa/repositories/TrackResourceRepository.java
@@ -4,5 +4,5 @@ import com.dovendev.track.jpa.entities.TracksResources;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-public interface TrackResourceRepository extends JpaRepository<TracksResources, Integer> {
+public interface TrackResourceRepository extends JpaRepository<TracksResources, Long> {
 }
